Complete set of parsing error messages
I complete the set of parsing error messages. There
were far too many of them for comfort, and it took several hours.
This is almost certainly because my grammar is large because there are (about) 36 infix operator symbols, and because there are several notations used to denote fnction expressions. There ought to be a better way of exploiting the context in the raw error messages files to do most of the generation automatically.