Constructors seem to be right
Next thing to explore is incorporating laziness properly.
Lazy abstractions work, but I think I want to be able to
declare some constructors to be lazy. Whether this
is a fixed (built-in) handful or a language feature remains to be
seen. Perhaps two kinds of Construct expression and two kinds of Cons value will be the easiest way.