Made the gamut of notation declarations consistent
outfix declarations with Nil and Cons constructors
read tuples of expressions then do a left reduce.
An example declaration is
   outfix {| |} Nil Cons
outfixes and leftfixes are quoted with dots

But there is the slight inconsistency that an outfix or a leftfix
goes into the environment as its leftmost (opening) symbol only, whereas it is quoted as specified above.

The machinery is not difficult to understand, but the
fact that there are so many operator symbols made it tedious to write all the necessary functions. (See, for example, mkExpr --
a workhorse used in mkOutfixList.

where their content should be, for example
 {|.|}
and
∀ . ∙ .